# public/openapi.yaml
openapi: 3.0.3
info:
  title: CRM+POS API
  description: RESTful API for SaaS CRM and Point of Sale System
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development
  - url: https://api.example.com/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        category_id: { type: integer }
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariant'
    ProductVariant:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        selling_price: { type: number }
        stock_qty: { type: integer }
...
paths:
...
  /products:
    get:
      tags: [Products]
      security: [{ bearerAuth: [] }]
      summary: List products
      parameters:
        - name: category_id
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }

    post:
      tags: [Products]
      security: [{ bearerAuth: [] }]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
  /products/{id}:
    get:
      tags: [Products]
      security: [{ bearerAuth: [] }]
      summary: Get product details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
...